Bug fixes and refactoring
